refactor: User 도메인과 서비스 계층 비밀번호 검증 방식 개선
