+++ src/main/kotlin/com/hoppingmall/mall/user/domain/User.kt
@@ -33,10 +33,6 @@ class User private constructor(
         }
     }
 
-    fun isPasswordMatch(rawPassword: String): Boolean {
-        return this.password.matches(rawPassword)
-    }
-
     fun updatePassword(newPassword: Password) {
         this.password = newPassword
     }
